Fix world projection numeric conversions

+++ Game/Rendering/WorldProjector.swift
@@ -14,10 +14,10 @@ final class WorldProjector {
         }
 
         let worldRect = CGRect(
-            x: layout.bounds.minX,
-            y: layout.bounds.minY,
-            width: layout.bounds.maxX - layout.bounds.minX,
-            height: layout.bounds.maxY - layout.bounds.minY
+            x: CGFloat(layout.bounds.minX),
+            y: CGFloat(layout.bounds.minY),
+            width: CGFloat(layout.bounds.maxX - layout.bounds.minX),
+            height: CGFloat(layout.bounds.maxY - layout.bounds.minY)
         )
         let asphalt = SKShapeNode(rect: worldRect)
         asphalt.fillColor = SKColor(white: 0.12, alpha: 1)
@@ -26,9 +26,9 @@ final class WorldProjector {
         root.addChild(asphalt)
 
         for obstacle in layout.obstacles {
-            let size = CGSize(width: obstacle.halfSize.x * 2, height: obstacle.halfSize.y * 2)
+            let size = CGSize(width: CGFloat(obstacle.halfSize.x * 2), height: CGFloat(obstacle.halfSize.y * 2))
             let node = SKShapeNode(rectOf: size, cornerRadius: 12)
-            node.position = CGPoint(x: obstacle.center.x, y: obstacle.center.y)
+            node.position = CGPoint(x: CGFloat(obstacle.center.x), y: CGFloat(obstacle.center.y))
             node.fillColor = SKColor(white: 0.24, alpha: 1)
             node.strokeColor = .systemYellow.withAlphaComponent(0.55)
             node.lineWidth = 3
@@ -42,7 +42,7 @@ final class WorldProjector {
     private func addParkingLines(to root: SKNode, bounds: WorldBounds) {
         for x in stride(from: bounds.minX + 90, through: bounds.maxX - 90, by: 90) {
             let line = SKShapeNode(rectOf: CGSize(width: 3, height: 72))
-            line.position = CGPoint(x: x, y: 0)
+            line.position = CGPoint(x: CGFloat(x), y: 0)
             line.fillColor = .white.withAlphaComponent(0.24)
             line.strokeColor = .clear
             root.addChild(line)
